web/requirements/development.txt
Django==2.2.13
djangorestframework==3.11.0
pytz==2019.3
gunicorn==20.0.4
python-magic==0.4.15
pytest==5.4.3
pytest-django==3.9.0
black==19.10b0
